DO NOT MERGE [CTS] Add a test for key/value backup and restore.
This test is intended to replace BackupTest in CTS Verifier (removed in
this change).

The test installs the app that generates random values and saves them in
different shared preferences and files. It uses BackupAgentHelper to do
key/value backup of those values. The tests verifies that the values are
restored after the app is uninstalled and reinstalled.

Test: make cts -j40 && cts-tradefed run cts -m CtsBackupHostTestCases
Bug: 32797101
Bug: 38492377
Change-Id: I580907c4ceb2cafb6d2372fbc05f6d8a9ac70e29
(cherry picked from commit 8b6f771a4c31902ab644396a161e269b69a747fb)