Nfc Verifier: Edit error message
HCE reader test is a 2 step process. First time the the devices are
tapped together the Emulator device will pop the option for picking the
technology TechA or TechB. And then tapping again would complete the
HCE test. On the reader side, the first tap will always throw an error
before the technology is selected. This patch edits the error message
such that it is not misleading.

Test: Run CtsVerifier test
Bug: 38396074
Change-Id: I9a13aacd673abd05334f5f947c202704e7bd9f30
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>