Fix: CTS Verifier app crashes Fingerprint Bound keys Test
The CTS Verifier app crashes due to using a non static
member class as fragment. It is impossible to recover the
state of the fragment upon configuration change. This was
never a problem for the Fingerprint Bound Keys Test because
it handles configuration changes and is not recreated.
However, as of SDK version N_MR1 a check was introduced
that checks the condition before it leads to problems
causing this crash.

This patch makes the FingerprintAuthDialogFragment a static
member class to fix the issue.

Also makes the
managedprovisioning.AuthenticationBoundKeyTestActivity
handle config events. Thost would crash the test as well.

Bug: 63066872
Test: Install and run CtsVerifier then perform the
      "Fingerprint Bound Keys Test"

Change-Id: I07aa9936cbb39efc9ea1942eb2758a0cb4c6919e