Fix CtsSignatureTestCases test failure in WindowManager
The WindowManager.getCurrentImeTouchRegion() is marked as @SystemApi so
is not in the current.txt file but is present in the runtime class which
caused a test failure. This updates the white list of methods to ignore
in interfaces, also updates the comment.

Bug: 63601982
Bug: 62194867
Bug: 35192974
Test: cts-tradefed run cts -m CtsSignatureTestCases
Change-Id: I6310d78b1d3acb18ba80872da3f974a38ce1978f
(cherry picked from commit dbdcb03398b1f49067a8aee5898c86b221cd1d8c)