cts-tradefed: Don't add . (working directory) to classpath.
The cts-tradefed script concatenates the list of jars to
pass to a java invocation by concatenating a list of paths,
separated by ':', into ${JAR_PATH}. The concatenation logic
precedes each element, including the first, by a ':'. This
led to the resulting classpath starting with ':', which
java interprets as the local directory being at the start
of the classpath.

This meant that if one had ever (accidentally) unzipped a
.jar file in the current working directory, the classes from
that .jar would henceforth occur at the start of the classpath,
which meant that they overrode classes occurring in the .jar
files, leading to hard-to-diagnose errors at cts-tradefed
runtime.

This CL adds logic to cut off the leading ':', meaning that
the current working directory is no longer prepended to the
classpath.

Test: A locally added "echo ${JAR_PATH}" in the cts-tradefed
      script confirms that that string no longer starts with ':'
Test: "cts-tradefed" no longer crashes when a stub version of
      org/json/JSONObject.class exists below the current
      working directory.

Change-Id: I3a6ef8307594eb3ab6096a9f858b8c6fd286af94