Disable Phone VR CTSV-tests on Standalone Headset.
The current CTS-Verifier VR tests are only meant to be run on phones and
only checks for VrListener life cycle. These checks will not work for
Standalone Headsets because headsets go through a different VrListener
life cycle. We should disable these tests in this situation for
standalone headsets.

Bug: 67782560
Test: CTS builds, verifier can launch, VR tests no longer present in the
list of tests. Tests are still present on sailfish.

Change-Id: I315ed671cc46ed8af634c5909b3fcadd9e12790b