Get view positions from one layout result
In testPromptViewAbove, the prompt view's position is retrieved before
WidgetTestUtils.runOnMainAndDrawSync, and the position of 1st child of
the list is retrieved after that. Those two position may come from two
different layout result, thus failed the requirement of the tests.

The change move the retrieval of prompt view's position after
WidgetTestUtils.runOnMainAndDrawSync. And the retrieval of the two
view's position happens in Ui thread, to guarantee they come from the
same layout result.

Bug 200360703

Test: CTS test

Change-Id: Iba0c8a7e733ca4719ae79c499464c752969985fd